# backend/ibutsu_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: A system to store and query test results
  title: Ibutsu API
  version: 1.13.2
servers:
  - url: /api
...
security:
  - jwt: []
paths:
  /result:
    get:
      description: |
        The `filter` parameter takes a list of filters to apply in the form of:

            {name}{operator}{value}

        where:

          - `name` is any valid column in the database
          - `operator` is one of `=`, `!`, `＞`, `＜`, `)`, `(`, `~`, `*`
          - `value` is what you want to filter by

        Operators are simple correspondents to MongoDB's query selectors:

          - `=` becomes `$eq`
          - `!` becomes `$ne`
          - `＞` becomes `$gt`
          - `＜` becomes `$lt`
          - `)` becomes `$gte`
          - `(` becomes `$lte`
          - `~` becomes `$regex`
          - `*` becomes `$in`
          - `@` becomes `$exists`

        Notes:

        - For the `$exists` operator, "true", "t", "yes", "y" and `1` will all be considered true,
          all other values are considered false.

        Example queries:

            /result?filter=metadata.run=63fe5
            /result?filter=test_id~neg
            /result?filter=result!passed
      operationId: get_result_list
      parameters:
        - description: Fields to filter by
          explode: true
          in: query
          name: filter
          required: false
          schema:
            items:
              type: string
            type: array
          style: form
        - description: Return an estimated count
          in: query
          name: estimate
          required: false
          schema:
            type: boolean
          style: form
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultList'
          description: successful operation
        201:
          content:
            application/json:
              schema:
                type: object
          description: Query being evaluated in a celery task
      summary: Get the list of results.
      tags:
        - result
      x-openapi-router-controller: ibutsu_server.controllers.result_controller
    post:
      operationId: add_result
      requestBody:
        $ref: '#/components/requestBodies/Result'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          description: Created
        400:
          description: Bad request, JSON required or not enough parameters
      summary: Create a test result
      tags:
        - result
      x-openapi-router-controller: ibutsu_server.controllers.result_controller
  /result/{id}:
    get:
      operationId: get_result
      parameters:
        - description: ID of result to return (uuid required)
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          description: Result item
        404:
          description: Result not found
      summary: Get a single result
      tags:
        - result
      x-openapi-router-controller: ibutsu_server.controllers.result_controller
    put:
      operationId: update_result
      parameters:
        - description: ID of result to update
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          style: simple
      requestBody:
        $ref: '#/components/requestBodies/Result'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          description: successful operation
        400:
          description: Bad reqest, JSON required or not enough parameters
        404:
          description: Result not found
      summary: Updates a single result
      tags:
        - result
      x-openapi-router-controller: ibutsu_server.controllers.result_controller
...
  /run:
    get:
      description: |
        The `filter` parameter takes a list of filters to apply in the form of:


            {name}{operator}{value}


        where:

          - `name` is any valid column in the database
          - `operator` is one of `=`, `!`, `＞`, `＜`, `)`, `(`, `~`, `*`
          - `value` is what you want to filter by

        Operators are simple correspondents to MongoDB's query selectors:

          - `=` becomes `$eq`
          - `!` becomes `$ne`
          - `＞` becomes `$gt`
          - `＜` becomes `$lt`
          - `)` becomes `$gte`
          - `(` becomes `$lte`
          - `~` becomes `$regex`
          - `*` becomes `$in`
          - `@` becomes `$exists`

        Notes:

        - For the `$exists` operator, "true", "t", "yes", "y" and `1` will all be considered true,
          all other values are considered false.

        Example queries:

            /run?filter=metadata.jenkins.job_name=jenkins_job
            /run?filter=summary.failures>0

      operationId: get_run_list
      parameters:
        - description: Fields to filter by
          explode: true
          in: query
          name: filter
          required: false
          schema:
            items:
              type: string
            type: array
          style: form
        - description: Return an estimated count
          in: query
          name: estimate
          required: false
          schema:
            type: boolean
          style: form
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
          description: Array of Runs
        201:
          content:
            application/json:
              schema:
                type: object
          description: Query being evaluated in a celery task
      summary: Get a list of the test runs
      tags:
        - run
      x-openapi-router-controller: ibutsu_server.controllers.run_controller
    post:
      operationId: add_run
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        description: Run item
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          description: A Run was created
        400:
          description: Bad request, JSON required
      summary: Create a run
      tags:
        - run
      x-openapi-router-controller: ibutsu_server.controllers.run_controller
  /runs/bulk-update:
    post:
      operationId: bulk_update
      parameters:
        - description: Fields to filter by
          explode: true
          in: query
          name: filter
          required: false
          schema:
            items:
              type: string
            type: array
          style: form
        - $ref: '#/components/parameters/PageSize'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRun'
        description: The metadata to add to the test runs
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
          description: Updated Run
        400:
          description: Bad request, JSON required
        401:
          description: Bad request, can only update metadata
        404:
          description: Filter(s) returned no Runs
        405:
          description: Bad request, cannot update more than 25 runs at a time
      summary: Update multiple runs with common metadata
      tags:
        - run
      x-openapi-router-controller: ibutsu_server.controllers.run_controller
  /run/{id}:
    get:
      operationId: get_run
      parameters:
        - description: ID of test run
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          description: Run item
        404:
          description: Run not found
      summary: Get a single run by ID (uuid required)
      tags:
        - run
      x-openapi-router-controller: ibutsu_server.controllers.run_controller
    put:
      operationId: update_run
      parameters:
        - description: ID of the test run
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        description: The updated test run
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          description: Updated Run
        400:
          description: Bad request, JSON required
        404:
          description: Run not found
      summary: Update a single run
      tags:
        - run
      x-openapi-router-controller: ibutsu_server.controllers.run_controller
  /project:
    get:
      operationId: get_project_list
      parameters:
        - description: Filter projects by owner ID
          explode: true
          in: query
          name: ownerId
          required: false
          schema:
            type: string
          style: form
        - description: Filter projects by group ID
          explode: true
          in: query
          name: groupId
          required: false
          schema:
            type: string
          style: form
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
          description: Array of projects
      summary: Get a list of projects
      tags:
        - project
      x-openapi-router-controller: ibutsu_server.controllers.project_controller
    post:
      operationId: add_project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        description: Project
        required: true
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: A project was created
        400:
          description: Bad request, JSON required
      summary: Create a project
      tags:
        - project
      x-openapi-router-controller: ibutsu_server.controllers.project_controller
...
components:
  requestBodies:
    Result:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
      description: Result item
...
  schemas:
    Result:
      example:
        id: a16ad60e-bf23-4195-99dc-594858ad3e5e
        duration: 6.027456183070403
        result: passed
        project_id: 9212ac64-348f-4c1c-90f7-fd0a45bcb47c
        component: login
        env: qa
        metadata:
          jenkins:
            build_number: 123
            job_name: test-jenkins-job
          project: 'insights-qe'
          commit_hash: F4BA3E12
        start_time: 2020-05-15T16:18:32.014053
        source: source
        params:
          provider: vmware
          ip_stack: ipv4
        test_id: test_click_on_button
      properties:
        id:
          description: Unique ID of the test result
          example: a16ad60e-bf23-4195-99dc-594858ad3e5e
          type: string
        test_id:
          description: Unique id
          example: test_click_on_button
          type: string
        start_time:
          description: Timestamp of starttime.
          type: string
        duration:
          description: Duration of test in seconds.
          type: number
        result:
          description: Status of result.
          enum:
          - passed
          - failed
          - error
          - skipped
          - xpassed
          - xfailed
          type: string
        component:
          description: A component
          example: login
          type: string
          nullable: true
        env:
          description: The environment which is being tested
          example: qa
          type: string
          nullable: true
        run_id:
          description: The run this result is associated with
          example: 64c2ab9e-cd64-4815-bf73-83b00c2e650f
          type: string
          nullable: true
        project_id:
          description: The project this run is associated with
          example: 44941c55-9736-42f6-acce-ca3c4739d0f3
          type: string
          nullable: true
        metadata:
          example:
            jenkins:
              job_name: test-jenkins-job
              build_number: 123
            project: insights-qe
            commit_hash: F4BA3E12
          type: object
        params:
          example:
            provider: vmware
            ip_stack: ipv4
          type: object
        source:
          description: Where the data came from (useful for filtering)
          type: string
      type: object
    Run:
      example:
        id: 64c2ab9e-cd64-4815-bf73-83b00c2e650f
        created: 2020-05-15T16:18:32.014053
        duration: 540.05433
        source: my-tests
        start_time: 2020-05-15T16:18:32.014053
        component: login
        env: qa
        summary:
          errors: 1
          failures: 3
          skips: 0
          xfailures: 0
          xpasses: 1
          tests: 548
        metadata:
          component: login
          env: qa
      properties:
        id:
          description: Unique ID of the test run
          example: 64c2ab9e-cd64-4815-bf73-83b00c2e650f
          type: string
        created:
          description: The time this record was created
          example: 2020-05-15T16:18:32.014053
          type: string
        duration:
          description: Duration of tests in seconds
          example: 154.778524555
          type: number
        source:
          description: A source for this test run
          example: my-tests
          type: string
          nullable: true
        start_time:
          description: The time the test run started
          example: 2020-05-15T16:18:32.014053
          type: string
        component:
          description: A component
          example: login
          type: string
          nullable: true
        env:
          description: The environment which is being tested
          example: qa
          type: string
          nullable: true
        project_id:
          description: The project this run is associated with
          example: 44941c55-9736-42f6-acce-ca3c4739d0f3
          type: string
          nullable: true
        summary:
          description: A summary of the test results
          type: object
        metadata:
          description: Extra metadata for this run
          type: object
          nullable: true
      type: object
    UpdateRun:
      example:
        metadata:
          annotations:
            - name: deploymentAnnotation
              value: Application was deployed at <timestamp>
              description: Optional description
            - name: otherAnnotation
              value: A different type of annotation
              description: Optional description
      properties:
        metadata:
          description: Extra data for this run
          type: object
      type: object
...
    Project:
      example:
        id: 44941c55-9736-42f6-acce-ca3c4739d0f3
        name: my-project
        title: My Project
        owner_id: 6b8b01ad-a17e-4ca1-8df5-fadb41439567
        group_id: a16ad60e-bf23-4195-99dc-594858ad3e5e
      properties:
        id:
          description: Unique ID of the project
          example: 44941c55-9736-42f6-acce-ca3c4739d0f3
          type: string
        name:
          description: The machine name of the project
          example: my-project
          type: string
        title:
          description: The human-readable title of the project
          example: My project
          type: string
        owner_id:
          description: The ID of the owner of this project
          example: 6b8b01ad-a17e-4ca1-8df5-fadb41439567
          type: string
        group_id:
          description: The ID of the group of this project
          example: a16ad60e-bf23-4195-99dc-594858ad3e5e
          type: string
      type: object
...
    Pagination:
      example:
        page: 2
        pageSize: 25
        totalPages: 10
        totalItems: 243
      properties:
        page:
          description: The current page number
          example: 2
          type: integer
        pageSize:
          description: The number of items per page
          example: 25
          type: integer
        totalPages:
          description: The total number of pages
          example: 10
          type: integer
        totalItems:
          description: The total number of items for this query
          example: 243
          type: integer
      type: object
    ResultList:
      example:
        results:
        - duration: 6.027456183070403
          result: passed
          metadata:
            jenkins_build: 145
            commit_hash: F4BA3E12
          starttime: 0.8008281904610115
          source: source
          params:
            provider: vmware
            ip_stack: ipv4
          test_id: test_id
        pagination:
          page: 2
          pageSize: 25
          totalPages: 10
          totalItems: 243
      properties:
        results:
          items:
            $ref: '#/components/schemas/Result'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
    RunList:
      example:
        runs:
        - id: 64c2ab9e-cd64-4815-bf73-83b00c2e650f
          duration: 540.05433
          summary:
            errors: 1
            failures: 3
            skips: 0
            xfailures: 0
            xpasses: 1
            tests: 548
        pagination:
          page: 2
          pageSize: 25
          totalPages: 10
          totalItems: 243
      properties:
        runs:
          items:
            $ref: '#/components/schemas/Run'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
...
    ProjectList:
      example:
        projects:
        - id: 44941c55-9736-42f6-acce-ca3c4739d0f3
          name: My Project
          ownerId: 6b8b01ad-a17e-4ca1-8df5-fadb41439567
          groupId: a16ad60e-bf23-4195-99dc-594858ad3e5e
        pagination:
          page: 2
          pageSize: 25
          totalPages: 10
          totalItems: 243
      properties:
        projects:
          items:
            $ref: '#/components/schemas/Project'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
...
  parameters:
    Page:
      description: Set the page of items to return, defaults to 1
      explode: true
      in: query
      name: page
      required: false
      schema:
        format: int32
        type: integer
      style: form
    PageSize:
      description: Set the number of items per page, defaults to 25
      explode: true
      in: query
      name: pageSize
      required: false
      schema:
        format: int32
        type: integer
      style: form
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: ibutsu_server.util.jwt.decode_token